more cleanup and instructions
