+++ package.json
@@ -1,17 +1,15 @@
 {
-    "name": "fragdenstaat_de",
+    "name": "uipa_org",
     "version": "0.0.0",
-    "description": "Freedom of Information Portal in Germany",
+    "description": "Freedom of Information Portal for Hawaii",
     "private": true,
     "scripts": {
       "test": "echo \"Error: no test specified\" && exit 1",
-      "lint": "eslint frontend --ext .js,.ts,.vue",
       "serve": "vite",
       "dev": "vite",
-      "build": "vite build",
-      "favicon": "real-favicon generate fragdenstaat_de/theme/static/img/favicon/favicon.json favicon-result.json fragdenstaat_de/theme/static/img/favicon"
+      "build": "vite build"
     },
-    "homepage": "https://github.com/okfde/fragdenstaat_de#readme",
+    "homepage": "https://github.com/codewithaloha/uipa",
     "devDependencies": {
       "@types/bootstrap": "^5.2.0",
       "@types/glidejs__glide": "^3.4.1",
