--- uipa_org/local_settings.py.example
@@ -1,4 +1,0 @@
-from .settings import ThemeBase
-
-class Dev(ThemeBase):
-    pass
